requirements.txt
biopython==1.84
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.53.1
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
lazypredict==0.2.12
lightgbm==4.5.0
matplotlib==3.9.1
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==23.2.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.1
scipy==1.13.1
seqfold==0.7.18
setuptools==68.2.0
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.5
tzdata==2024.1
wheel==0.41.2
xgboost==2.1.1
zipp==3.19.2
umap-learn